backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.135.1"}
psycopg = {extras = ["binary"], version = "^3.3.3"}
pyjwt = "^2.11.0"
bcrypt = "^5.0.0"
python-dotenv = "^1.2.2"
aiosqlite = "^0.22.1"
openai = "^2.24.0"
apscheduler = "^3.11.2"
httpx = "^0.28.1"
cryptography = "^44.0.0"
playwright = "^1.58.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
